split utils in 2 different files in utils folder
