+++ packages/space-header/src/utils/get_space.ts
@@ -1,0 +1,12 @@
+import type { Space } from "./../type";
+
+export const get_space = async (space_id: string): Promise<Space | null> => {
+	try {
+		const response = await fetch(`https://huggingface.co/api/spaces/${space_id}`);
+		const data = await response.json();
+		console.log(data);
+		return data as Space;
+	} catch (error) {
+		return null;
+	}
+};
